Move histo tests and helpers to seperate files
